requirements.txt
aniso8601==3.0.0
click==6.7
Flask==0.12.2
flask-marshmallow==0.8.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.15.0
marshmallow-sqlalchemy==0.13.2
pytz==2018.3
six==1.11.0
SQLAlchemy==1.2.5
Werkzeug==0.14.1
pyyaml
